#105 Renamed ui to snapshots
