+++ core/src/main/java/com/github/glusk2/sprouts/core/snapshots/MoveDrawing.java
@@ -1,4 +1,4 @@
-package com.github.glusk2.sprouts.core.ui;
+package com.github.glusk2.sprouts.core.snapshots;
 
 import java.util.ArrayList;
 import java.util.Iterator;
