+++ core/src/main/java/com/github/glusk2/sprouts/core/snapshots/package-info.java
@@ -2,4 +2,4 @@
  * Contains classes and interfaces that are used to create and display the
  * user interface for a game of Sprouts.
  */
-package com.github.glusk2.sprouts.core.ui;
+package com.github.glusk2.sprouts.core.snapshots;
