+++ core/src/test/java/com/github/glusk2/sprouts/core/snapshots/package-info.java
@@ -1,0 +1,2 @@
+/** UI test package. */
+package com.github.glusk2.sprouts.core.snapshots;
