--- core/src/test/java/com/github/glusk2/sprouts/core/ui/package-info.java
@@ -1,2 +1,0 @@
-/** UI test package. */
-package com.github.glusk2.sprouts.core.ui;
